ai-memory-system/requirements.txt
# MCP SDK
mcp>=0.9.0

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# Vector Database
qdrant-client>=1.7.0

# HTTP Client
httpx>=0.25.2

# Ollama
ollama>=0.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
